+++ cosmetics-web/app/controllers/responsible_persons/notification_files_controller.rb
@@ -1,6 +1,5 @@
 class ResponsiblePersons::NotificationFilesController < ApplicationController
   before_action :set_responsible_person
-  before_action :set_notification_file
 
   def new; end
 
@@ -9,20 +8,36 @@ class ResponsiblePersons::NotificationFilesController < ApplicationController
   end
 
   def create
-    unless notification_file_params && notification_file_params[:uploaded_file]
-      @notification_file.errors.add :uploaded_file, "No file selected"
+    @errors = []
+    if uploaded_files_params.nil?
+      @errors << { text: "No files selected", href: "#uploaded_files" }
       return render :new
     end
 
-    @notification_file.name = notification_file_params[:uploaded_file].original_filename
-    @notification_file.responsible_person = @responsible_person
-    @notification_file.user = User.current
-
-    if @notification_file.save
-      redirect_to responsible_person_notifications_path(@responsible_person)
-    else
-      render :new
+    if uploaded_files_params.length > NotificationFile.get_max_number_of_files
+      @errors << {
+          text: "Too many files selected. Please select no more than #{NotificationFile.get_max_number_of_files} files",
+          href: "#uploaded_files"
+      }
+      return render :new
     end
+
+    uploaded_files_params.each do |uploaded_file|
+      notification_file = NotificationFile.new(
+        name: uploaded_file.original_filename,
+        responsible_person: @responsible_person,
+        user: User.current
+      )
+      notification_file.uploaded_file.attach(uploaded_file)
+
+      unless notification_file.save
+        @errors.concat(notification_file.errors.full_messages.map { |message|
+          { text: message, href: "#file-upload-form-group" }
+        })
+        return render :new
+      end
+    end
+    redirect_to responsible_person_notifications_path(@responsible_person)
   end
 
   def destroy
@@ -38,19 +53,15 @@ class ResponsiblePersons::NotificationFilesController < ApplicationController
 private
 
   # Use callbacks to share common setup or constraints between actions.
-  def set_notification_file
-    @notification_file = NotificationFile.new(notification_file_params)
-  end
-
   def set_responsible_person
     @responsible_person = ResponsiblePerson.find(params[:responsible_person_id])
     authorize @responsible_person, :show?
   end
 
   # Never trust parameters from the scary internet, only allow the white list through.
-  def notification_file_params
-    if params.has_key?(:notification_file)
-      params.require(:notification_file).permit(:name, :uploaded_file)
+  def uploaded_files_params
+    if params.has_key?(:uploaded_files)
+      params.require(:uploaded_files)
     end
   end
 end