+++ cosmetics-web/spec/factories/notification.rb
@@ -1,8 +1,15 @@
 FactoryBot.define do
   factory :notification do
+    sequence(:product_name) { |n| "Product #{n}" }
+
     factory :draft_notification do
       state { :draft_complete }
     end
+
+    factory :imported_notification do
+      state { :notification_file_imported }
+    end
+
     factory :registered_notification do
       state { :notification_complete }
     end