+++ cosmetics-web/spec/system/manual_notification_entry_spec.rb
@@ -1,9 +1,11 @@
 require 'rails_helper'
 
 RSpec.describe "Manually enter product details", type: :system do
+  let(:responsible_person) { create(:responsible_person) }
+
   before do
-    sign_in_as_member_of_responsible_person(create(:responsible_person))
     mock_antivirus_api
+    sign_in_as_member_of_responsible_person(responsible_person)
   end
 
   after do
@@ -12,7 +14,7 @@ RSpec.describe "Manually enter product details", type: :system do
   end
 
   it "allows user to complete notification" do
-    visit new_notification_path
+    visit new_responsible_person_notification_path(responsible_person)
 
     # add_product_name
     fill_in :notification_product_name, with: "Super Shampoo"
@@ -31,15 +33,11 @@ RSpec.describe "Manually enter product details", type: :system do
     click_button "Continue"
 
     # add_product_image
-    attach_file(
-      :image_upload,
-      Rails.root + 'spec/fixtures/testImage.png'
-)
+    attach_file(:image_upload, Rails.root + 'spec/fixtures/testImage.png')
     click_button "Continue"
 
-    mark_images_as_safe
-
     # Check your answers page
+    notification = get_notification_from_edit_page_url
     expect_check_your_answers_value("Product name", "Super Shampoo")
     expect_check_your_answers_value("Imported", "No")
     expect_check_your_answers_value("Number of components", "1")
@@ -48,12 +46,40 @@ RSpec.describe "Manually enter product details", type: :system do
     click_button "Accept and register the cosmetics product"
 
     # Check notification was completed
-    notification = get_notification_from_url
-    expect(notification.state).to eq("notification_complete")
+    expect(notification.reload.state).to eq("notification_complete")
+  end
+
+  it "associates responsible person with notification" do
+    visit new_responsible_person_notification_path(responsible_person)
+
+    # add_product_name
+    fill_in :notification_product_name, with: "Super Shampoo"
+    click_button "Continue"
+
+    # is_imported
+    choose("No")
+    click_button "Continue"
+
+    # single_or_multi_component
+    choose("Yes")
+    click_button "Continue"
+
+    # number_of_shades
+    choose("No")
+    click_button "Continue"
+
+    # add_product_image
+    attach_file(:image_upload, Rails.root + 'spec/fixtures/testImage.png')
+    click_button "Continue"
+
+    # Check your answers page
+    expect_check_your_answers_value("Name", responsible_person.name)
+    expect_check_your_answers_value("Email address", responsible_person.email_address)
+    expect_check_your_answers_value("Phone number", responsible_person.phone_number)
   end
 
   it "allows user to complete notification for imported cosmetics" do
-    visit new_notification_path
+    visit new_responsible_person_notification_path(responsible_person)
 
     # add_product_name
     fill_in :notification_product_name, with: "Super Shampoo"
@@ -76,15 +102,11 @@ RSpec.describe "Manually enter product details", type: :system do
     click_button "Continue"
 
     # add_product_image
-    attach_file(
-      :image_upload,
-      Rails.root + 'spec/fixtures/testImage.png'
-)
+    attach_file(:image_upload, Rails.root + 'spec/fixtures/testImage.png')
     click_button "Continue"
 
-    mark_images_as_safe
-
     # Check your answers page
+    notification = get_notification_from_edit_page_url
     expect_check_your_answers_value("Product name", "Super Shampoo")
     expect_check_your_answers_value("Imported", "Yes")
     expect_check_your_answers_value("Imported from", "New Zealand")
@@ -94,12 +116,11 @@ RSpec.describe "Manually enter product details", type: :system do
     click_button "Accept and register the cosmetics product"
 
     # Check notification was completed
-    notification = get_notification_from_url
-    expect(notification.state).to eq("notification_complete")
+    expect(notification.reload.state).to eq("notification_complete")
   end
 
   it "allows user to complete notification for cosmetics with multiple shades" do
-    visit new_notification_path
+    visit new_responsible_person_notification_path(responsible_person)
 
     # add_product_name
     fill_in :notification_product_name, with: "Super Shampoo"
@@ -127,15 +148,11 @@ RSpec.describe "Manually enter product details", type: :system do
     click_button "Continue"
 
     # add_product_image
-    attach_file(
-      :image_upload,
-      Rails.root + 'spec/fixtures/testImage.png'
-)
+    attach_file(:image_upload, Rails.root + 'spec/fixtures/testImage.png')
     click_button "Continue"
 
-    mark_images_as_safe
-
     # Check your answers page
+    notification = get_notification_from_edit_page_url
     expect_check_your_answers_value("Product name", "Super Shampoo")
     expect_check_your_answers_value("Imported", "No")
     expect_check_your_answers_value("Number of components", "1")
@@ -145,8 +162,7 @@ RSpec.describe "Manually enter product details", type: :system do
     click_button "Accept and register the cosmetics product"
 
     # Check notification was completed
-    notification = get_notification_from_url
-    expect(notification.state).to eq("notification_complete")
+    expect(notification.reload.state).to eq("notification_complete")
   end
 
 private
@@ -156,23 +172,11 @@ private
     expect(row).to have_text(value)
   end
 
-  def get_notification_from_url
-    if (match = current_url.match(%r!/notifications/(\d+)/!))
+  def get_notification_from_edit_page_url
+    if (match = current_url.match(%r!/notifications/(\d+)/edit!))
       reference_number = match.captures[0].to_i
     end
 
     Notification.find_by reference_number: reference_number
-  end
-
-  # The worker doesn't mark system test images as safe, so we have to do it
-  # manually to allow the manual journey to finish.
-  def mark_images_as_safe
-    notification = get_notification_from_url
-
-    notification.image_uploads.each do |image_upload|
-      blob = image_upload.file.blob
-      blob.metadata = { safe: true }
-      blob.save
-    end
   end
 end