+++ mspsds-web/app/models/concerns/searchable.rb
@@ -3,64 +3,7 @@ module Searchable
   extend ActiveSupport::Concern
 
   included do
-    include Elasticsearch::Model
-    include Elasticsearch::Model::Callbacks
-
-    # The following dynamic templates define custom mappings for the major data types
-    # that automatically generate appropriate sort fields for each type.
-    settings do
-      mapping dynamic_templates: [
-        {
-          strings: {
-            match_mapping_type: "string",
-            mapping: {
-              type: "text",
-              fields: {
-                sort: {
-                  type: "icu_collation_keyword"
-                }
-              }
-            }
-          }
-        }, {
-          numbers: {
-            match_mapping_type: "long",
-            mapping: {
-              "type": "long",
-              fields: {
-                sort: {
-                  type: "long"
-                }
-              }
-            }
-          }
-        }, {
-          dates: {
-            match_mapping_type: "date",
-            mapping: {
-              type: "date",
-              fields: {
-                sort: {
-                  type: "date"
-                }
-              }
-            }
-          }
-        }, {
-          booleans: {
-            match_mapping_type: "boolean",
-            mapping: {
-              type: "boolean",
-              fields: {
-                sort: {
-                  type: "boolean"
-                }
-              }
-            }
-          }
-        }
-      ]
-    end
+    include Shared::Web::Concerns::Searchable
 
     # TODO: These lines fix the issue of not updating the updated_at in Elasticsearch.
     # Same issue is pointed out in the following link. We can remove it once that PR is merged.