+++ shared-web/app/clients/shared/web/keycloak_client.rb
@@ -36,7 +36,10 @@ module Shared
       include Singleton
 
       def initialize
+        # The gem we're using has its api split across these three classes
         @client = Keycloak::Client
+        @admin = Keycloak::Admin
+        @internal = Keycloak::Internal
         super
       end
 
@@ -108,8 +111,9 @@ module Shared
         JSON.parse(response)
       end
 
-      def registration_url
-        Keycloak::Client.auth_server_url + "/realms/#{Keycloak::Client.realm}/protocol/openid-connect/registrations?client_id=#{Keycloak::Client.client_id}&response_type=code"
+      def registration_url(redirect_uri)
+        params = URI.encode_www_form(client_id: Keycloak::Client.client_id, response_type: "code", redirect_uri: redirect_uri)
+        Keycloak::Client.auth_server_url + "/realms/#{Keycloak::Client.realm}/protocol/openid-connect/registrations?#{params}"
       end
 
       def login_url(redirect_uri)
@@ -142,8 +146,13 @@ module Shared
         { id: user["sub"], email: user["email"], groups: user["groups"], first_name: user["given_name"], last_name: user["family_name"] }
       end
 
-      def has_role?(role)
-        @client.has_role? role
+      def has_role?(user_id, role)
+        if User.current && (User.current.id == user_id)
+          # This is faster, as it uses the already fetched claims
+          @client.has_role? role
+        else
+          @internal.has_role? user_id, role
+        end
       end
 
     private